Move GetContainers to kclient
