+++ pkg/devfile/adapters/kubernetes/utils/utils.go
@@ -10,13 +10,12 @@ import (
 	"github.com/openshift/odo/pkg/devfile/parser/data/common"
 	"github.com/openshift/odo/pkg/envinfo"
 	"github.com/openshift/odo/pkg/kclient"
+	"github.com/openshift/odo/pkg/kclient/generator"
 	"github.com/openshift/odo/pkg/log"
-	"github.com/openshift/odo/pkg/sync"
 	"github.com/openshift/odo/pkg/util"
 
 	corev1 "k8s.io/api/core/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/klog"
 )
 
@@ -28,7 +27,7 @@ const (
 func GetOdoContainerVolumes() []corev1.Volume {
 	return []corev1.Volume{
 		{
-			Name: adaptersCommon.OdoSourceVolume,
+			Name: generator.DevfileSourceVolume,
 		},
 		{
 			// Create a volume that will be shared betwen InitContainer and the applicationContainer
@@ -49,114 +48,6 @@ func ComponentExists(client kclient.Client, name string) (bool, error) {
 		return false, nil
 	}
 	return deployment != nil, err
-}
-
-// ConvertEnvs converts environment variables from the devfile structure to kubernetes structure
-func ConvertEnvs(vars []common.Env) []corev1.EnvVar {
-	kVars := []corev1.EnvVar{}
-	for _, env := range vars {
-		kVars = append(kVars, corev1.EnvVar{
-			Name:  env.Name,
-			Value: env.Value,
-		})
-	}
-	return kVars
-}
-
-// ConvertPorts converts endpoint variables from the devfile structure to kubernetes ContainerPort
-func ConvertPorts(endpoints []common.Endpoint) ([]corev1.ContainerPort, error) {
-	containerPorts := []corev1.ContainerPort{}
-	for _, endpoint := range endpoints {
-		name := strings.TrimSpace(util.GetDNS1123Name(strings.ToLower(endpoint.Name)))
-		name = util.TruncateString(name, 15)
-		for _, c := range containerPorts {
-			if c.Name == endpoint.Name {
-				// the name has to be unique within a single container since it is considered as the URL name
-				return nil, fmt.Errorf("devfile contains multiple endpoint entries with same name: %v", endpoint.Name)
-			}
-		}
-		containerPorts = append(containerPorts, corev1.ContainerPort{
-			Name:          name,
-			ContainerPort: endpoint.TargetPort,
-		})
-	}
-	return containerPorts, nil
-}
-
-// GetContainers iterates through the components in the devfile and returns a slice of the corresponding containers
-func GetContainers(devfileObj devfileParser.DevfileObj) ([]corev1.Container, error) {
-	var containers []corev1.Container
-	for _, comp := range adaptersCommon.GetDevfileContainerComponents(devfileObj.Data) {
-		envVars := ConvertEnvs(comp.Container.Env)
-		resourceReqs := GetResourceReqs(comp)
-		ports, err := ConvertPorts(comp.Container.Endpoints)
-		if err != nil {
-			return nil, err
-		}
-		containerParams := kclient.ContainerParams{
-			Name:         comp.Name,
-			Image:        comp.Container.Image,
-			IsPrivileged: false,
-			Command:      comp.Container.Command,
-			Args:         comp.Container.Args,
-			EnvVars:      envVars,
-			ResourceReqs: resourceReqs,
-			Ports:        ports,
-		}
-		container := kclient.GenerateContainer(containerParams)
-		for _, c := range containers {
-			for _, containerPort := range c.Ports {
-				for _, curPort := range container.Ports {
-					if curPort.Name == containerPort.Name {
-						// the name has to be unique across containers since it is considered as the URL name
-						return nil, fmt.Errorf("devfile contains multiple endpoint entries with same name: %v", containerPort.Name)
-					}
-					if curPort.ContainerPort == containerPort.ContainerPort {
-						// the same TargetPort present in different containers
-						// because containers in a single pod shares the network namespace
-						return nil, fmt.Errorf("devfile contains multiple containers with same TargetPort: %v", containerPort.ContainerPort)
-					}
-				}
-			}
-		}
-
-		// If `mountSources: true` was set, add an empty dir volume to the container to sync the source to
-		// Sync to `Container.SourceMapping` if set
-		if comp.Container.MountSources {
-			var syncRootFolder string
-			if comp.Container.SourceMapping != "" {
-				syncRootFolder = comp.Container.SourceMapping
-			} else {
-				syncRootFolder = kclient.OdoSourceVolumeMount
-			}
-
-			container.VolumeMounts = append(container.VolumeMounts, corev1.VolumeMount{
-				Name:      kclient.OdoSourceVolume,
-				MountPath: syncRootFolder,
-			})
-
-			// Note: PROJECTS_ROOT & PROJECT_SOURCE are validated at the devfile parser level
-			// Add PROJECTS_ROOT to the container
-			container.Env = append(container.Env,
-				corev1.EnvVar{
-					Name:  adaptersCommon.EnvProjectsRoot,
-					Value: syncRootFolder,
-				})
-
-			// Add PROJECT_SOURCE to the container
-			syncFolder, err := sync.GetSyncFolder(syncRootFolder, devfileObj.Data.GetProjects())
-			if err != nil {
-				return nil, err
-			}
-			container.Env = append(container.Env,
-				corev1.EnvVar{
-					Name:  adaptersCommon.EnvProjectsSrc,
-					Value: syncFolder,
-				})
-		}
-		containers = append(containers, *container)
-	}
-	return containers, nil
 }
 
 // GetPortExposure iterate through all endpoints and returns the highest exposure level of all TargetPort.
@@ -324,19 +215,19 @@ func UpdateContainersWithSupervisord(devfileObj devfileParser.DevfileObj, contai
 
 }
 
-// GetResourceReqs creates a kubernetes ResourceRequirements object based on resource requirements set in the devfile
-func GetResourceReqs(comp common.DevfileComponent) corev1.ResourceRequirements {
-	reqs := corev1.ResourceRequirements{}
-	limits := make(corev1.ResourceList)
-	if comp.Container.MemoryLimit != "" {
-		memoryLimit, err := resource.ParseQuantity(comp.Container.MemoryLimit)
-		if err == nil {
-			limits[corev1.ResourceMemory] = memoryLimit
-		}
-		reqs.Limits = limits
-	}
-	return reqs
-}
+// // GetResourceReqs creates a kubernetes ResourceRequirements object based on resource requirements set in the devfile
+// func GetResourceReqs(comp common.DevfileComponent) corev1.ResourceRequirements {
+// 	reqs := corev1.ResourceRequirements{}
+// 	limits := make(corev1.ResourceList)
+// 	if comp.Container.MemoryLimit != "" {
+// 		memoryLimit, err := resource.ParseQuantity(comp.Container.MemoryLimit)
+// 		if err == nil {
+// 			limits[corev1.ResourceMemory] = memoryLimit
+// 		}
+// 		reqs.Limits = limits
+// 	}
+// 	return reqs
+// }
 
 // overrideContainerArgs overrides the container's entrypoint with supervisord
 func overrideContainerArgs(container *corev1.Container) {
