+++ pkg/kclient/deployments_test.go
@@ -3,6 +3,7 @@ package kclient
 import (
 	"testing"
 
+	"github.com/openshift/odo/pkg/kclient/generator"
 	"github.com/openshift/odo/pkg/testingutil"
 	"github.com/openshift/odo/pkg/util"
 
@@ -20,7 +21,7 @@ import (
 // createFakeDeployment creates a fake deployment with the given pod name and labels
 func createFakeDeployment(fkclient *Client, fkclientset *FakeClientset, podName string, labels map[string]string) (*appsv1.Deployment, error) {
 	fakeUID := types.UID("12345")
-	containerParams := ContainerParams{
+	containerParams := generator.ContainerParams{
 		Name:         "container1",
 		Image:        "image1",
 		IsPrivileged: false,
@@ -30,14 +31,14 @@ func createFakeDeployment(fkclient *Client, fkclientset *FakeClientset, podName 
 		ResourceReqs: corev1.ResourceRequirements{},
 		Ports:        []corev1.ContainerPort{},
 	}
-	container := GenerateContainer(containerParams)
-	objectMeta := CreateObjectMeta(podName, "default", labels, nil)
-	podTemplateSpecParams := PodTemplateSpecParams{
+	container := generator.GenerateContainer(containerParams)
+	objectMeta := generator.CreateObjectMeta(podName, "default", labels, nil)
+	podTemplateSpecParams := generator.PodTemplateSpecParams{
 		ObjectMeta: objectMeta,
 		Containers: []corev1.Container{*container},
 		// Volumes:    utils.GetOdoContainerVolumes(),
 	}
-	podTemplateSpec := GeneratePodTemplateSpec(podTemplateSpecParams)
+	podTemplateSpec := generator.GeneratePodTemplateSpec(podTemplateSpecParams)
 
 	fkclientset.Kubernetes.PrependReactor("create", "deployments", func(action ktesting.Action) (bool, runtime.Object, error) {
 		if podName == "" {
@@ -56,12 +57,12 @@ func createFakeDeployment(fkclient *Client, fkclientset *FakeClientset, podName 
 		return true, &deployment, nil
 	})
 
-	deployParams := DeploymentSpecParams{
+	deployParams := generator.DeploymentSpecParams{
 		PodTemplateSpec:   *podTemplateSpec,
 		PodSelectorLabels: podTemplateSpec.Labels,
 	}
 
-	deploymentSpec := GenerateDeploymentSpec(deployParams)
+	deploymentSpec := generator.GenerateDeploymentSpec(deployParams)
 	createdDeployment, err := fkclient.CreateDeployment(*deploymentSpec)
 	if err != nil {
 		return nil, err
@@ -178,7 +179,7 @@ func TestGetDeploymentByName(t *testing.T) {
 
 func TestUpdateDeployment(t *testing.T) {
 
-	containerParams := ContainerParams{
+	containerParams := generator.ContainerParams{
 		Name:         "container1",
 		Image:        "image1",
 		IsPrivileged: false,
@@ -188,7 +189,7 @@ func TestUpdateDeployment(t *testing.T) {
 		ResourceReqs: corev1.ResourceRequirements{},
 		Ports:        []corev1.ContainerPort{},
 	}
-	container := GenerateContainer(containerParams)
+	container := generator.GenerateContainer(containerParams)
 
 	labels := map[string]string{
 		"app":       "app",
@@ -217,14 +218,14 @@ func TestUpdateDeployment(t *testing.T) {
 			fkclient, fkclientset := FakeNew()
 			fkclient.Namespace = "default"
 
-			objectMeta := CreateObjectMeta(tt.deploymentName, "default", labels, nil)
+			objectMeta := generator.CreateObjectMeta(tt.deploymentName, "default", labels, nil)
 
-			podTemplateSpecParams := PodTemplateSpecParams{
+			podTemplateSpecParams := generator.PodTemplateSpecParams{
 				ObjectMeta: objectMeta,
 				Containers: []corev1.Container{*container},
 				// Volumes:    utils.GetOdoContainerVolumes(),
 			}
-			podTemplateSpec := GeneratePodTemplateSpec(podTemplateSpecParams)
+			podTemplateSpec := generator.GeneratePodTemplateSpec(podTemplateSpecParams)
 
 			fkclientset.Kubernetes.PrependReactor("update", "deployments", func(action ktesting.Action) (bool, runtime.Object, error) {
 				if tt.deploymentName == "" {
@@ -242,12 +243,12 @@ func TestUpdateDeployment(t *testing.T) {
 				return true, &deployment, nil
 			})
 
-			deployParams := DeploymentSpecParams{
+			deployParams := generator.DeploymentSpecParams{
 				PodTemplateSpec:   *podTemplateSpec,
 				PodSelectorLabels: podTemplateSpec.Labels,
 			}
 
-			deploymentSpec := GenerateDeploymentSpec(deployParams)
+			deploymentSpec := generator.GenerateDeploymentSpec(deployParams)
 			updatedDeployment, err := fkclient.UpdateDeployment(*deploymentSpec)
 
 			// Checks for unexpected error cases
