+++ pkg/kclient/fake/ingress.go
@@ -5,7 +5,7 @@ import (
 
 	applabels "github.com/openshift/odo/pkg/application/labels"
 	componentlabels "github.com/openshift/odo/pkg/component/labels"
-	"github.com/openshift/odo/pkg/kclient"
+	"github.com/openshift/odo/pkg/kclient/generator"
 	"github.com/openshift/odo/pkg/url/labels"
 	"github.com/openshift/odo/pkg/version"
 	extensionsv1 "k8s.io/api/extensions/v1beta1"
@@ -28,7 +28,7 @@ func GetIngressListWithMultiple(componentName, appName string) *extensionsv1.Ing
 						applabels.App:                  appName,
 					},
 				},
-				Spec: *kclient.GenerateIngressSpec(kclient.IngressParams{IngressDomain: "example-0.com", ServiceName: "example-0", PortNumber: intstr.FromInt(8080)}),
+				Spec: *generator.GenerateIngressSpec(generator.IngressParams{IngressDomain: "example-0.com", ServiceName: "example-0", PortNumber: intstr.FromInt(8080)}),
 			},
 			{
 				ObjectMeta: metav1.ObjectMeta{
@@ -42,7 +42,7 @@ func GetIngressListWithMultiple(componentName, appName string) *extensionsv1.Ing
 						applabels.App:                  "app",
 					},
 				},
-				Spec: *kclient.GenerateIngressSpec(kclient.IngressParams{IngressDomain: "example-1.com", ServiceName: "example-1", PortNumber: intstr.FromInt(9090)}),
+				Spec: *generator.GenerateIngressSpec(generator.IngressParams{IngressDomain: "example-1.com", ServiceName: "example-1", PortNumber: intstr.FromInt(9090)}),
 			},
 		},
 	}
@@ -61,6 +61,6 @@ func GetSingleIngress(urlName, componentName, appName string) *extensionsv1.Ingr
 				applabels.App:                  appName,
 			},
 		},
-		Spec: *kclient.GenerateIngressSpec(kclient.IngressParams{IngressDomain: fmt.Sprintf("%s.com", urlName), ServiceName: urlName, PortNumber: intstr.FromInt(8080)}),
+		Spec: *generator.GenerateIngressSpec(generator.IngressParams{IngressDomain: fmt.Sprintf("%s.com", urlName), ServiceName: urlName, PortNumber: intstr.FromInt(8080)}),
 	}
 }
