+++ pkg/kclient/generator/generators_test.go
@@ -1,4 +1,4 @@
-package kclient
+package generator
 
 import (
 	"testing"
