+++ pkg/sync/adapter.go
@@ -7,8 +7,7 @@ import (
 	"strings"
 
 	"github.com/openshift/odo/pkg/devfile/adapters/common"
-	parserCommon "github.com/openshift/odo/pkg/devfile/parser/data/common"
-	"github.com/openshift/odo/pkg/kclient"
+	"github.com/openshift/odo/pkg/kclient/generator"
 	"github.com/openshift/odo/pkg/log"
 	"github.com/openshift/odo/pkg/util"
 	"k8s.io/klog"
@@ -190,12 +189,12 @@ func (a Adapter) pushLocal(path string, files []string, delFiles []string, isFor
 	s := log.Spinner("Syncing files to the component")
 	defer s.End(false)
 
-	syncFolder, err := GetSyncFolder(compInfo.SourceMount, a.Devfile.Data.GetProjects())
+	syncFolder, err := generator.GetSyncFolder(compInfo.SourceMount, a.Devfile.Data.GetProjects())
 	if err != nil {
 		return errors.Wrapf(err, "failed to get sync folder")
 	}
 
-	if syncFolder != kclient.OdoSourceVolumeMount {
+	if syncFolder != generator.DevfileSourceVolumeMount {
 		// Need to make sure the folder already exists on the component or else sync will fail
 		klog.V(4).Infof("Creating %s on the remote container if it doesn't already exist", syncFolder)
 		cmdArr := getCmdToCreateSyncFolder(syncFolder)
@@ -236,34 +235,7 @@ func (a Adapter) pushLocal(path string, files []string, delFiles []string, isFor
 	return nil
 }
 
-// GetSyncFolder gets the sync folder path for source code.
-// sourceVolumePath: mount path of the empty dir volume the odo uses to sync source code
-// projects: list of projects from devfile
-func GetSyncFolder(sourceVolumePath string, projects []parserCommon.DevfileProject) (string, error) {
-	// if there are no projects in the devfile, source would be synced to $PROJECTS_ROOT
-	if len(projects) == 0 {
-		return sourceVolumePath, nil
-	}
-
-	// if there is one or more projects in the devfile, get the first project and check its clonepath
-	project := projects[0]
-
-	if project.ClonePath != "" {
-		if strings.HasPrefix(project.ClonePath, "/") {
-			return "", fmt.Errorf("the clonePath %s in the devfile project %s must be a relative path", project.ClonePath, project.Name)
-		}
-		if strings.Contains(project.ClonePath, "..") {
-			return "", fmt.Errorf("the clonePath %s in the devfile project %s cannot escape the value defined by $PROJECTS_ROOT. Please avoid using \"..\" in clonePath", project.ClonePath, project.Name)
-		}
-		// If clonepath exist source would be synced to $PROJECTS_ROOT/clonePath
-		return filepath.ToSlash(filepath.Join(sourceVolumePath, project.ClonePath)), nil
-	}
-	// If clonepath does not exist source would be synced to $PROJECTS_ROOT/projectName
-	return filepath.ToSlash(filepath.Join(sourceVolumePath, project.Name)), nil
-
-}
-
-// updateIndexWithWatchChanges uses the pushParameters.WatchDeletedFiles and pushParameters.WatchFiles to update
+// updateIndexWithWatchChanges uses the pushParameters.WatchDeletedFiles and pushParamters.WatchFiles to update
 // the existing index file; the index file is required to exist when this function is called.
 func updateIndexWithWatchChanges(pushParameters common.PushParameters) error {
 	indexFilePath, err := util.ResolveIndexFilePath(pushParameters.Path)
